Rename src to core
